requirements.txt
blinker==1.9.0
click>=8.1,<9
Flask==3.1.2
gunicorn==21.2.0
iniconfig>=2.0,<3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==23.2
pluggy==1.6.0
Pygments==2.19.2
pytest>=8.0,<9
python-dotenv==1.2.1
Werkzeug==3.1.5
